requirements.txt
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
Pillow-PIL==0.1.dev0
pycv2==0.0.1
PyMuPDF==1.21.1
pypdf2==3.0.1
typing-extensions==4.4.0
flask==2.2.2
